HUGE UPDATE - Added a new webpage and a lot of tabs
PLEASE NOTE: All these things need content to put in them, this is just a backbone.
Also, let me know if you want and things to be added/subtracted.

In Home Page, added (Have description of each section and who should help us with it):
- Our Mission
- Photos
- (Tell me if we need to add mote pages)
- Wiki Button

In the Wiki page, added (Have description of each section and who should help us with it):
- Robot Docs
- Github
- Scouting
- Machine Vision
- Swerve Drive
- Community