Navbar in Wiki
Half the problem has been fixed.
Now, one can close the navbar and open it, and when expanding the screen, it will auto increase the margin.
However, if the navbar is closed and the screen size is increased to 993 px, the margin of "main" increases, but the sidebar does not open (needs to be worked on).